fix(mcp): break tools.ts ↔ toolSearch.ts cycle via leaf toolDefinition.ts
The #5269 merge extracted the tool_search defs into schemas/toolSearch.ts, which
imports McpToolDefinition from tools.ts while tools.ts imports toolSearchTool from
toolSearch.ts — a cycle that fails check:cycles on release/v3.8.40. Move the shared
AuditLevel + McpToolDefinition types into a leaf (toolDefinition.ts) that both import;
tools.ts re-exports them for backward compatibility.
