+++ open-sse/mcp-server/schemas/toolSearch.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import type { McpToolDefinition } from "./tools.ts";
+import type { McpToolDefinition } from "./toolDefinition.ts";
 
 export const toolSearchInput = z
   .object({
